containers/prepro/requirements.txt
numpy == 1.21.3
Pillow == 8.4.0

